+++ example/methods/account.js
@@ -12,6 +12,7 @@ exports.test1 = function(data) {
 //  console.log('redis Expire time', this.redis());
   //console.log(arguments);
   deferred.resolve('Yeahhhh');
+  console.log('TYPES =>', this.Types);
   // test redis
   /*this.redis().instance.add('aaa', { foo : 'abar' }, this.redis().expire);
   this.redis().instance.get('aaa').then(function (success) {